requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
distro==1.9.0
EbookLib==0.18
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.1
openai==1.65.5
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
